Re #INTRD-24735 Fixed invoice validation

+++ opencell-admin/ejbs/src/main/java/org/meveo/service/payments/impl/AccountOperationService.java
@@ -858,8 +858,8 @@ public class AccountOperationService extends PersistenceService<AccountOperation
 
     public void fillOperationNumber(AccountOperation accountOperation)
     {
-        BigInteger operationNumber =(BigInteger) getEntityManager().createNativeQuery("select nextval('account_operation_number_seq')").getSingleResult();
-        accountOperation.setOperationNumber(operationNumber.longValue());
+        Long operationNumber = (Long) getEntityManager().createNativeQuery("select nextval('account_operation_number_seq')").getSingleResult();
+        accountOperation.setOperationNumber(operationNumber);
     }
 
     @Override
